feat : merge encryptLocalService and encryptSessionService to encryptStorage
